+++ cs/src/core/Utilities/SpanByte/SpanByte.cs
@@ -94,79 +94,4 @@ namespace FASTER.core
             Buffer.MemoryCopy(Unsafe.AsPointer(ref this), Unsafe.AsPointer(ref dst), fullLength, fullLength);
         }
     }
-
-    /// <summary>
-    /// Equality comparer for SpanByte
-    /// </summary>
-    public struct SpanByteComparer : IFasterEqualityComparer<SpanByte>
-    {
-        /// <inheritdoc />
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public unsafe long GetHashCode64(ref SpanByte spanByte)
-        {
-            byte* ptr = (byte*)Unsafe.AsPointer(ref spanByte);
-            return Utility.HashBytes(ptr, (*(int*)ptr) + sizeof(int));
-        }
-
-        /// <inheritdoc />
-        /// <inheritdoc />
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public unsafe bool Equals(ref SpanByte k1, ref SpanByte k2)
-        {
-            return
-                new ReadOnlySpan<byte>(Unsafe.AsPointer(ref k1), k1.length + sizeof(int))
-                    .SequenceEqual(
-                new ReadOnlySpan<byte>(Unsafe.AsPointer(ref k2), k2.length + sizeof(int)));
-        }
-    }
-
-    /// <summary>
-    /// IVariableLengthStruct for SpanByte
-    /// </summary>
-    public struct SpanByteLength : IVariableLengthStruct<SpanByte>
-    {
-        /// <inheritdoc />
-        public int GetInitialLength()
-        {
-            return sizeof(int);
-        }
-
-        /// <inheritdoc />
-        public int GetLength(ref SpanByte t)
-        {
-            return sizeof(int) + t.length;
-        }
-    }
-
-    /// <summary>
-    /// Callback functions for SpanByte
-    /// </summary>
-    public class SpanByteFunctions<Context> : FunctionsBase<SpanByte, SpanByte, byte[], byte[], Context>
-    {
-        /// <inheritdoc />
-        public override void SingleReader(ref SpanByte key, ref byte[] input, ref SpanByte value, ref byte[] dst)
-        {
-            dst = value.ToByteArray();
-        }
-
-        /// <inheritdoc />
-        public override void ConcurrentReader(ref SpanByte key, ref byte[] input, ref SpanByte value, ref byte[] dst)
-        {
-            dst = value.ToByteArray();
-        }
-
-        /// <inheritdoc />
-        public override void SingleWriter(ref SpanByte key, ref SpanByte src, ref SpanByte dst)
-        {
-            src.CopyTo(ref dst);
-        }
-
-        /// <inheritdoc />
-        public override bool ConcurrentWriter(ref SpanByte key, ref SpanByte src, ref SpanByte dst)
-        {
-            if (dst.length < src.length) return false;
-            src.CopyTo(ref dst);
-            return true;
-        }
-    }
 }
